*Ajout de nouveaux médias pour l'interface (médias de tests) *Le système d'event sur les boutons est totalement fonctionnel : Le callback permet de récupérer l'id du bouton cliqué sur la fenêtre MyWin. *Implémentation du système de label. Il est maintenant possible d'écrire du texte dans une fenêtre *Changement de l'écriture XML des fenêtres : L'élément "controls" à été changé par "window" et ce dernier contients les attributs de la fenêtre (position, image de fond, ...) *Ajout d'un attribut alpha pour permettre de rendre la fenêtre transparente
/!\ Les widgets ne dépendent pas encore de la fenêtre. Plus tard leurs positions seront calculés en fonction de la position de la fenêtre mère pour plus de clarté.
Il est possible d'ajouté un système de déplacement de la fenêtre mais il n'est pas prioritaire.
A terme le système de fenêtre remplacera le HUD actuel, et permettra l'accès aux options ainsi qu'au menu principal.

+++ trunk/src/gui_system/Control.hpp
@@ -11,7 +11,7 @@ namespace gui{
         * Typedef
         */
         typedef unsigned int ControlID;
-        typedef sf::Vector2i ControlPos;
+        typedef sf::Vector2f ControlPos;
 
         Control();
         Control(ControlID id, ControlPos pos);
